Added parser for FBO nightly files
The parser collects all of the information for the following announcement types:
	Presolicitation (PRESOL)
	Combined Synopsis/Solicitation (COMBINE)
	Sources Sought (SRCSGT)
	Special Notice (SNOTE)
These are the types that seem to be relevant to our users.
The most recent Nightly XML file is downloaded using the XML_download.php
script, then each field is constructed into an SQL statement and sent to the
database.

It currently does not add interest tags.