Finished FBO weekly parser. Adds and updates records.
It reads through all of the records in the XML file one by one, adding each
record or updating existing records. The assumption here is that the weekly
file takes precedence over anything else. So if a record in the XML file
contradicts a record in our database, our record will be overwritten.

Also, the records are kind of odd. Some of them don't have solution numbers.
Others have strange date formats, and the descriptions have HTML tags in them.
The script rejects records with no solution numbers, and ignores improper
dates.